Include the prebuilt 0.28.4 libraries for Linux and Mac OS X.
In addition, don't use pkgconfig to drive the C compiler and
linker. This is to avoid the need to have the full platform-specific
libgit2 installation available, either in tree or elsewhere.

The go dummies are required in the header directories. If they are not
there the golang vendoring won't bring in the header files and they
are required to compile the go code with cgo imports.